requirements.txt
torch>=2.1.2
torchvision>=0.12.0
timm>=0.9.12
einops>=0.7.0
rasterio>=1.3.9
lightning>=2.1.3
wandb>=0.16.1
optuna>=3.5.0
sentinelhub
overpy
osmnx
geopandas